Transport

GenTransportMCMC.jl uses an optimized version of the Trace Transform DSL to encode pushforward distributions. Users may perform the transport map accelerated inference algorithms using either predefined families of maps $\{ \tilde T(\cdot; \gamma) \}_{\gamma \in\mathcal A}$ or those of their own devising, using the Transport API. This API is centered around the following assumptions of the family, which are discussed in the Mathematical Details section.

There is a fixed integer $d \in \mathbb N$ such that each $\tilde T(\cdot; \gamma): \mathbb R^d \rightarrow \mathbb R^d$.

Each map $\tilde T(\cdot; \gamma)$ is invertible.

Each map $\tilde T(\cdot; \gamma)$ has a lower-triangular Jacobian.

The parameter set $\mathcal A$ for the family satisfies $\mathcal A = \prod_{i=1}^d \mathbb R^{\mathcal J_i}$ for finite sets $\mathcal J_i$ with accompanying collections of functions $(\psi_j)_{j\in\mathcal J_i}$.

For each $\gamma = (\gamma_1, \ldots, \gamma_d) \in \mathcal A$, the $i$-th component $\tilde T_i(\cdot; \gamma_i)$ of the map $\tilde T(\cdot; \gamma)$ is determined as such.

\[T_i(\theta; \gamma_i) = \sum_{j\in\mathcal J_i} \gamma_{i,j} \psi_j(\theta)\]

Transport API

GenTransportMCMC.OrderedBasis — Type

abstract type OrderedBasis end

Abstract type corresponding to an ordered basis of scalar-functions $(ψ_j)_{j\in\mathcal J_i}$. An instance tuple of this type will act as a map, so that tuple(θ) will correspond to the vector $(ψ_j(θ))_{j \in \mathcal J_i}$.

To implement a subtype B <: OrderedBasis, one must implement:

(tuple::B)(θ::AbstractArray{<:Real})::AbstractArray{<:Real}
Base.length(tuple::B)::Int64

GenTransportMCMC.TransportFamily — Type

T = TransportFamily(bases::AbstractArray{<:OrderedBasis})

Provide an array bases of OrderedBasis instances where each element bases[i] corresponds to some ordered basis $(ψ_j)_{j\in\mathcal J_i}$ and receive the induced family $\{ \tilde T(\cdot; γ) \}_{γ \in \mathcal A}$ (T).

With an instance T of a TransportFamily, one may apply T(γ) to a parameter γ and get a TransportMap instance corresponding to the map $T(\cdot; γ)$. Recall, this is defined so that the $i$-th component $T_i(\cdot; γ_i)$ is as follows.

\[T_i(θ; γ_i) = \sum_{j\in\mathcal J_i} γ_{i,j} ψ_j(θ)\]

GenTransportMCMC.TransportMap — Type

map = TransportMap(dim, parameter, f)

Wrap a provided function f with a TransportMap type for the type checking. In addition, we keep track of the dimension (dim) of the vector space on which f is defined and the parameter (parameter) of the specific map.

GenTransportMCMC.inverse — Function

theta = inverse(tmap::TransportMap, r)

Return $θ$ such that $T(θ; γ) = r$, where $T(\cdot; γ)$ is our transport map (tmap).

GenTransportMCMC.sample_cost — Function

cost = sample_cost(tmap::TransportMap, samples::AbstractArray{<:AbstractArray{<:Real}})

Compute the cost $C(\tilde T(\cdot; γ))$ of a transport map $\tilde T(\cdot; γ)$ based off of samples $θ_1, \ldots, θ_K$.

\[C\big(\tilde T(\cdot; γ)\big) = \sum_{i=1}^d\sum_{k=1}^K \Big( \frac{1}{2} \tilde T_i^2(θ^{(k)}; γ_i) - \log\frac{∂\tilde T_i(\cdot; γ_i)}{∂θ_i}(θ^{(k)}) \Big)\]

Transport Families

GenTransportMCMC.hermite — Function

y = hermite(n::Int, x::Real)

Evaluate the $n$-degree Hermite polynomial at $x$.

\[n!\sum_{m=0}^{[n/2]} \frac{(-1)^m}{m!(n-2m)!} \frac{x^{n-2m}}{2^m}\]

GenTransportMCMC.HermiteBasis — Type

basis = HermiteBasis(vectors)

Encode an ordered basis (basis) of polynomials $(ψ_j)_{j\in \mathcal J_i}$ of the form

\[ψ_{(j_1,\ldots,j_n)}(θ_1, \ldots, θ_n) = \prod_{k=1}^n φ_{j_k}(θ_k),\]

where $φ_{j_k}$ is the Hermite polynomial of degree $j_k$. The indices $\mathcal J_i$ of the polynomials is provided by vectors in the natural way. For example,

HermiteBasis([[1, 2, 3], [0, 1, 1, 2]])

will encode the ordered basis $( ψ_{(1,2,3)}, ψ_{(0,1,1,2)} )$.